backend/requirements.txt
fastapi
uvicorn[standard]
tortoise-orm[asyncpg]
aerich
python-multipart
aiofiles
python-jose[cryptography]
passlib[bcrypt]
pydantic
aiohttp
numpy
GDAL==3.10.3
geopandas
pytest
pytest-asyncio
celery[redis]
redis
flower